refactor: handle adapter errors in `server/index.js`
