+++ src/server/index.js
@@ -2,7 +2,7 @@ import jwt from "../lib/jwt"
 import parseUrl from "../lib/parse-url"
 import logger, { setLogger } from "../lib/logger"
 import * as cookie from "./lib/cookie"
-import { withErrorHandling, defaultEvents } from "./lib/default-events"
+import { defaultEvents } from "./lib/default-events"
 import * as defaultCallbacks from "./lib/default-callbacks"
 import parseProviders from "./lib/providers"
 import * as routes from "./routes"
@@ -13,6 +13,7 @@ import extendRes from "./lib/extend-res"
 import csrfTokenHandler from "./lib/csrf-token-handler"
 import * as pkce from "./lib/oauth/pkce-handler"
 import * as state from "./lib/oauth/state-handler"
+import { eventsErrorHandler, adapterErrorHandler } from "src/lib/errors"
 
 // To work properly in production with OAuth providers the NEXTAUTH_URL
 // environment variable must be set.
@@ -124,10 +125,11 @@ async function NextAuthHandler(req, res, userOptions) {
         ...userOptions.jwt,
       },
       // Event messages
-      events: withErrorHandling(
+      events: eventsErrorHandler(
         { ...defaultEvents, ...userOptions.events },
         logger
       ),
+      adapter: adapterErrorHandler(userOptions.adapter, logger),
       // Callback functions
       callbacks: {
         ...defaultCallbacks,
