+++ src/server/lib/default-events.js
@@ -1,5 +1,3 @@
-import { upperSnake } from "../../lib/errors"
-
 /** @type {import("types").EventCallbacks} */
 export const defaultEvents = {
   signIn() {},
@@ -8,23 +6,4 @@ export const defaultEvents = {
   updateUser() {},
   linkAccount() {},
   session() {},
-}
-
-/**
- * Wraps an object of methods and adds error handling.
- * @param {import("types").EventCallbacks} methods
- * @param {import("types").LoggerInstance} logger
- * @return {import("types").EventCallbacks}
- */
-export function withErrorHandling(methods, logger) {
-  return Object.entries(methods).reduce((acc, [name, method]) => {
-    acc[name] = async (...args) => {
-      try {
-        return await method(...args)
-      } catch (e) {
-        logger.error(`${upperSnake(name)}_EVENT_ERROR`, e)
-      }
-    }
-    return acc
-  }, {})
 }
